+++ src/matrix_transformations/rotation.c
@@ -12,47 +12,47 @@
 
 #include "minirt.h"
 
-t_matrix *rotation_x(float radians)
+t_matrix *rotation_x(double radians)
 {
     t_matrix *result;
 
     result = create_identity(4);
     if(!result)
         return (NULL);
-    result->data[1][1] = cosf(radians);
-    result->data[2][2] = cosf(radians);
-    result->data[2][1] = sinf(radians);
-    result->data[1][2] = -sinf(radians);
+    result->data[1][1] = cos(radians);
+    result->data[2][2] = cos(radians);
+    result->data[2][1] = sin(radians);
+    result->data[1][2] = -sin(radians);
 
     return (result);
 }
 
-t_matrix *rotation_y(float radians)
+t_matrix *rotation_y(double radians)
 {
     t_matrix *result;
 
     result = create_identity(4);
     if(!result)
         return (NULL);
-    result->data[0][0] = cosf(radians);
-    result->data[0][2] = sinf(radians);
-    result->data[2][0] = -sinf(radians);
-    result->data[2][2] = cosf(radians);
+    result->data[0][0] = cos(radians);
+    result->data[0][2] = sin(radians);
+    result->data[2][0] = -sin(radians);
+    result->data[2][2] = cos(radians);
 
     return (result);
 }
 
-t_matrix *rotation_z(float radians)
+t_matrix *rotation_z(double radians)
 {
     t_matrix *result;
 
     result = create_identity(4);
     if(!result)
         return (NULL);
-    result->data[0][0] = cosf(radians);
-    result->data[0][1] = -sinf(radians);
-    result->data[1][0] = sinf(radians);
-    result->data[1][1] = cosf(radians);
+    result->data[0][0] = cos(radians);
+    result->data[0][1] = -sin(radians);
+    result->data[1][0] = sin(radians);
+    result->data[1][1] = cos(radians);
 
     return (result);
 }
