+++ src/tuples/tuple.c
@@ -12,17 +12,17 @@
 
 #include "minirt.h"
 
-t_tuple new_tuple(float x, float y, float z, float w)
+t_tuple new_tuple(double x, double y, double z, double w)
 {
     return ((t_tuple){x,y,z,w});
 }
 
-t_tuple new_point(float x, float y, float z)
+t_tuple new_point(double x, double y, double z)
 {
         return ((t_tuple){x,y,z,1});
 }
 
-t_tuple new_vector(float x, float y, float z)
+t_tuple new_vector(double x, double y, double z)
 {
         return ((t_tuple){x,y,z,0});
 }
